# online-city-tickets-backend/src/main/resources/static/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Online City Tickets - API Gateway",
        "version": "v1"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api/v1",
            "description": "Development Server"
        }
    ],
    "paths": {
...
        "/account": {
            "get": {
                "tags": [
                    "accounts"
                ],
                "summary": "Get current account's details.",
                "operationId": "getAccount",
                "responses": {
                    "200": {
                        "description": "Fetched current account's details successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AccountDto"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Attempted to get account of an admin."
                    },
                    "401": {
                        "description": "Unauthorized."
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "accounts"
                ],
                "summary": "Update current account's details.",
                "operationId": "updateAccount",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateAccountReq"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Account updated successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AccountDto"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request body or phone number was supplied for an inspector account or attempted to update an admin account."
                    },
                    "401": {
                        "description": "Unauthorized."
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/credit-cards": {
            "get": {
                "tags": [
                    "cards"
                ],
                "summary": "List all saved credit cards of the current account.",
                "operationId": "listCreditCards",
                "responses": {
                    "200": {
                        "description": "Fetched saved credit cards successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CreditCardDto"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized or used on a inspector account."
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "cards"
                ],
                "summary": "Save a new credit card on the current account.",
                "operationId": "saveCreditCard",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SaveCreditCardReq"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Credit card saved successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CreditCardDto"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request body or card is invalid/expired."
                    },
                    "401": {
                        "description": "Unauthorized or used on a inspector account."
                    },
                    "409": {
                        "description": "Credit card already saved."
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/credit-cards/{id}": {
            "get": {
                "tags": [
                    "cards"
                ],
                "summary": "Get details of a saved credit card.",
                "operationId": "getCreditCard",
                "responses": {
                    "200": {
                        "description": "Fetched credit card details successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CreditCardDto"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized or on a inspector account."
                    },
                    "404": {
                        "description": "Credit card not found."
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "cards"
                ],
                "summary": "Update details of a saved credit card.",
                "operationId": "updateCreditCard",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCreditCardReq"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Credit card updated successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CreditCardDto"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request body or card is invalid/expired."
                    },
                    "401": {
                        "description": "Unauthorized or on a inspector account."
                    },
                    "404": {
                        "description": "Credit card not found."
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "cards"
                ],
                "summary": "Delete a saved credit card.",
                "operationId": "deleteCreditCard",
                "responses": {
                    "204": {
                        "description": "Credit card deleted successfully."
                    },
                    "401": {
                        "description": "Unauthorized or on a inspector account."
                    },
                    "404": {
                        "description": "Credit card not found."
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            },
            "parameters": [
                {
                    "$ref": "#/components/parameters/cardId"
                }
            ]
        },
...
        "/tickets": {
            "get": {
                "tags": [
                    "tickets"
                ],
                "summary": "List all tickets of the current account.",
                "operationId": "listTickets",
                "responses": {
                    "200": {
                        "description": "Fetched tickets successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TicketDto"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized or on a inspector account."
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "tickets"
                ],
                "summary": "Purchase a new ticket.",
                "description": "This operation charges the ticket price from the virtual wallet.",
                "operationId": "purchaseTicket",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PurchaseTicketReq"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Ticket purchased successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TicketDto"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request body."
                    },
                    "401": {
                        "description": "Unauthorized or on a inspector account."
                    },
                    "402": {
                        "description": "Insufficient funds in virtual wallet."
                    },
                    "404": {
                        "description": "Ticket offer not found."
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/tickets/{code}": {
            "get": {
                "tags": [
                    "tickets"
                ],
                "summary": "Get details of a single ticket of the current account.",
                "operationId": "getTicket",
                "responses": {
                    "200": {
                        "description": "Fetched ticket successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TicketDto"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized or on a inspector account."
                    },
                    "404": {
                        "description": "Ticket not found."
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            },
            "parameters": [
                {
                    "$ref": "#/components/parameters/code"
                }
            ]
        },
...
    },
    "components": {
        "schemas": {
            "BaseAccountDto": {
                "required": [
                    "fullName",
                    "email"
                ],
                "type": "object",
                "properties": {
                    "fullName": {
                        "$ref": "#/components/schemas/fullName"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    }
                }
            },
            "PassengerDto": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/BaseAccountDto"
                    },
                    {
                        "required": [
                            "type",
                            "walletBalanceGrosze"
                        ],
                        "properties": {
                            "type": {
                                "type": "string",
                                "enum": [
                                    "passenger"
                                ]
                            },
                            "walletBalanceGrosze": {
                                "type": "integer",
                                "minimum": 0
                            },
                            "phoneNumber": {
                                "$ref": "#/components/schemas/phoneNumber"
                            }
                        }
                    }
                ]
            },
            "InspectorDto": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/BaseAccountDto"
                    },
                    {
                        "required": [
                            "type"
                        ],
                        "properties": {
                            "type": {
                                "type": "string",
                                "enum": [
                                    "inspector"
                                ]
                            }
                        }
                    }
                ]
            },
            "AccountDto": {
                "oneOf": [
                    {
                        "$ref": "#/components/schemas/PassengerDto"
                    },
                    {
                        "$ref": "#/components/schemas/InspectorDto"
                    }
                ]
            },
...
            "UpdateAccountReq": {
                "type": "object",
                "properties": {
                    "fullName": {
                        "$ref": "#/components/schemas/fullName"
                    },
                    "newPassword": {
                        "$ref": "#/components/schemas/password"
                    },
                    "phoneNumber": {
                        "$ref": "#/components/schemas/phoneNumber"
                    }
                }
            },
            "CreditCardDto": {
                "required": [
                    "id",
                    "lastFourDigits",
                    "holderName",
                    "expirationDate"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "$ref": "#/components/schemas/id"
                    },
                    "label": {
                        "type": "string",
                        "example": "My favourite card"
                    },
                    "lastFourDigits": {
                        "type": "string",
                        "pattern": "^[0-9]{4}$"
                    },
                    "holderName": {
                        "$ref": "#/components/schemas/fullName"
                    },
                    "expirationDate": {
                        "$ref": "#/components/schemas/expirationDate"
                    }
                }
            },
            "SaveCreditCardReq": {
                "required": [
                    "number",
                    "holderName",
                    "expirationDate"
                ],
                "type": "object",
                "properties": {
                    "label": {
                        "type": "string",
                        "example": "My favourite card"
                    },
                    "number": {
                        "$ref": "#/components/schemas/creditCardNumber"
                    },
                    "holderName": {
                        "$ref": "#/components/schemas/fullName"
                    },
                    "expirationDate": {
                        "$ref": "#/components/schemas/expirationDate"
                    }
                }
            },
            "UpdateCreditCardReq": {
                "type": "object",
                "properties": {
                    "label": {
                        "type": "string"
                    },
                    "expirationDate": {
                        "$ref": "#/components/schemas/expirationDate"
                    }
                }
            },
...
            "BaseTicketOfferDto": {
                "required": [
                    "id",
                    "displayNameEn",
                    "displayNamePl",
                    "kind",
                    "priceGrosze"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "$ref": "#/components/schemas/id"
                    },
                    "displayNameEn": {
                        "type": "string",
                        "example": "30-minute, reduced"
                    },
                    "displayNamePl": {
                        "type": "string",
                        "example": "30-minutowy, ulgowy"
                    },
                    "kind": {
                        "type": "string",
                        "enum": [
                            "standard",
                            "reduced"
                        ]
                    },
                    "priceGrosze": {
                        "type": "integer",
                        "minimum": 1,
                        "example": 320
                    }
                }
            },
            "SingleFareTicketOfferDto": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/BaseTicketOfferDto"
                    },
                    {
                        "required": [
                            "scope"
                        ],
                        "properties": {
                            "scope": {
                                "type": "string",
                                "enum": [
                                    "single-fare"
                                ]
                            }
                        }
                    }
                ]
            },
            "TimeLimitedTicketOfferDto": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/BaseTicketOfferDto"
                    },
                    {
                        "required": [
                            "scope",
                            "durationMinutes"
                        ],
                        "properties": {
                            "scope": {
                                "type": "string",
                                "enum": [
                                    "time-limited"
                                ]
                            },
                            "durationMinutes": {
                                "type": "integer",
                                "minimum": 1
                            }
                        }
                    }
                ]
            },
            "LongTermTicketOfferDto": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/BaseTicketOfferDto"
                    },
                    {
                        "required": [
                            "scope",
                            "intervalInDays"
                        ],
                        "properties": {
                            "scope": {
                                "type": "string",
                                "enum": [
                                    "long-term"
                                ]
                            },
                            "intervalInDays": {
                                "type": "integer",
                                "minimum": 1
                            }
                        }
                    }
                ]
            },
            "TicketOfferDto": {
                "oneOf": [
                    {
                        "$ref": "#/components/schemas/SingleFareTicketOfferDto"
                    },
                    {
                        "$ref": "#/components/schemas/TimeLimitedTicketOfferDto"
                    },
                    {
                        "$ref": "#/components/schemas/LongTermTicketOfferDto"
                    }
                ]
            },
            "PurchaseTicketReq": {
                "required": [
                    "offerId"
                ],
                "type": "object",
                "properties": {
                    "offerId": {
                        "$ref": "#/components/schemas/id"
                    }
                }
            },
            "TicketDto": {
                "required": [
                    "code",
                    "purchaseTime",
                    "offer"
                ],
                "type": "object",
                "properties": {
                    "code": {
                        "$ref": "#/components/schemas/ticketCode"
                    },
                    "purchaseTime": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "offer": {
                        "$ref": "#/components/schemas/TicketOfferDto"
                    },
                    "validation": {
                        "$ref": "#/components/schemas/ValidationDto"
                    }
                }
            },
...
            "ValidationDto": {
                "required": [
                    "time",
                    "vehicleSideNumber"
                ],
                "type": "object",
                "properties": {
                    "time": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "vehicleSideNumber": {
                        "$ref": "#/components/schemas/vehicleSideNumber"
                    }
                }
            },
...
            "id": {
                "type": "number",
                "minimum": 0,
                "example": 1
            },
...
            "fullName": {
                "type": "string",
                "example": "John Smith",
                "minLength": 1
            },
            "password": {
                "type": "string",
                "format": "password",
                "minLength": 8,
                "example": "password"
            },
            "phoneNumber": {
                "type": "string",
                "pattern": "^[0-9]{9}$"
            },
            "creditCardNumber": {
                "type": "string",
                "pattern": "^[0-9]{16}$"
            },
            "expirationDate": {
                "type": "string",
                "pattern": "^(0[1-9]|1[0-2])/[0-9]{2}$"
            },
...
            "ticketCode": {
                "type": "string",
                "pattern": "^[0-9A-Z]{10}$"
            },
            "vehicleSideNumber": {
                "type": "string",
                "pattern": "^[0-9]{4}$"
            },
...
        },
        "parameters": {
            "code": {
                "name": "code",
                "in": "path",
                "description": "Unique code of the ticket.",
                "required": true,
                "schema": {
                    "$ref": "#/components/schemas/ticketCode"
                },
                "example": "3K7G9T2E"
            },
            "cardId": {
                "name": "id",
                "in": "path",
                "description": "Unique identifier of the credit card.",
                "required": true,
                "schema": {
                    "pattern": "^(0|[1-9][0-9]*)$",
                    "type": "string"
                },
                "example": "1"
            },
...
        },
        "securitySchemes": {
            "jwt": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
}